Rating + wishlisting

+++ src/main/java/com/solo/kinocavern/dao/CategoryDAO.java
@@ -8,7 +8,7 @@ public interface CategoryDAO {
 
     public List<Category> findAll();
 
-    public Category findById(int id);
+    public Category findById(Long id);
 
     public void save(Category category);
