+++ src/main/java/com/solo/kinocavern/dao/CountryDAO.java
@@ -9,10 +9,10 @@ public interface CountryDAO {
 
     public List<Country> findAll();
 
-    public Country findById(int id);
+    public Country findById(Long id);
 
     public void save(Country country);
 
-    public List<Country> findByIds(List<Integer> ids);
+    public List<Country> findByIds(List<Long> ids);
 
 }
