+++ src/main/java/com/solo/kinocavern/entity/Country.java
@@ -11,7 +11,7 @@ public class Country {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name="id")
-    private int id;
+    private Long id;
 
     @Column(name="name")
     private String name;
@@ -30,16 +30,16 @@ public class Country {
     public Country() {
     }
 
-    public Country(int id, String name) {
+    public Country(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
